completed game

+++ games/24_points/eval.h
@@ -1,5 +1,5 @@
 #ifndef EVAL_24_POINTS_H
 #define EVAL_24_POINTS_H
 #include <stdbool.h>
-extern bool eval(char *, int, int, int, int, int *);
+extern bool eval(char *, int, int, int, int, double *);
 #endif
